# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  isManager: boolean;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface EmployeeWithBalance {
  id: number;
  name: string;
  role: string;
  department: string;
  hireDate: string;
  createdAt: string;
  /** Available vacation days (earned - taken) */
  vacationBalanceDays: number;
  /** Days until next vacation period starts (or 0 if vacation is now due) */
  daysUntilNextVacation: number;
  isOnVacation: boolean;
  /** @nullable */
  nextVacationStart: string | null;
}

export type VacationItemStatus =
  (typeof VacationItemStatus)[keyof typeof VacationItemStatus];

export const VacationItemStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface VacationItem {
  id: number;
  employeeId: number;
  startDate: string;
  endDate: string;
  durationDays: number;
  /** @nullable */
  notes: string | null;
  status: VacationItemStatus;
  createdAt: string;
}

export interface EmployeeDetail {
  id: number;
  name: string;
  role: string;
  department: string;
  hireDate: string;
  createdAt: string;
  vacationBalanceDays: number;
  daysUntilNextVacation: number;
  isOnVacation: boolean;
  /** @nullable */
  nextVacationStart: string | null;
  vacations: VacationItem[];
}

export interface ListEmployeesResponse {
  employees: EmployeeWithBalance[];
}

export interface MyEmployeeEnvelope {
  employee: EmployeeWithBalance | null;
}

export interface RegisterEmployeeBody {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  role: string;
  /** @minLength 1 */
  department: string;
  hireDate: string;
}

export interface CreateEmployeeBody {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  role: string;
  /** @minLength 1 */
  department: string;
  hireDate: string;
}

export interface UpdateEmployeeBody {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  role?: string;
  /** @minLength 1 */
  department?: string;
  hireDate?: string;
}

export interface ListVacationsResponse {
  vacations: VacationItem[];
}

export type UpdateVacationStatusBodyStatus =
  (typeof UpdateVacationStatusBodyStatus)[keyof typeof UpdateVacationStatusBodyStatus];

export const UpdateVacationStatusBodyStatus = {
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdateVacationStatusBody {
  status: UpdateVacationStatusBodyStatus;
}

export interface CreateVacationBody {
  startDate: string;
  endDate: string;
  /** @nullable */
  notes?: string | null;
}

export interface UpcomingVacation {
  employeeId: number;
  employeeName: string;
  department: string;
  startDate: string;
  endDate: string;
  durationDays: number;
}

export type DashboardSummaryDepartmentBreakdownItem = {
  department: string;
  count: number;
  onVacation: number;
};

export interface DashboardSummary {
  totalEmployees: number;
  onVacationToday: number;
  upcomingVacations: UpcomingVacation[];
  /** Employees with overdue vacation balance (>= 30 days balance and no upcoming vacation) */
  overdueVacations: number;
  departmentBreakdown: DashboardSummaryDepartmentBreakdownItem[];
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};