# lib/api-zod/src/generated/types/vacationItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VacationItemStatus } from "./vacationItemStatus";

export interface VacationItem {
  id: number;
  employeeId: number;
  startDate: Date;
  endDate: Date;
  durationDays: number;
  /** @nullable */
  notes: string | null;
  status: VacationItemStatus;
  createdAt: Date;
}